final version of data preparation

+++ prepare_datasets/prepare_splits.py
@@ -1,6 +1,7 @@
 import pandas as pd
 import os
 import kmeans
+import shutil
 
 from sklearn.preprocessing import LabelEncoder
 from sklearn.model_selection import train_test_split
@@ -8,7 +9,6 @@ from matplotlib import pyplot as plt
 
 annotation_datasets = '../annotated_datasets/CSV/'
 columns =['Image Filename','coordinates','Term']
-dst = '/path/to/your/dataset'
 
 def get_full_dataset():
     """
@@ -21,19 +21,60 @@ def get_full_dataset():
             they can't be part of our training-testing splits.
     """
 
-    flickr = pd.read_csv(annotation_datasets + 'Flickr/full_dataset_Flickr.csv', error_bad_lines=False)
-    kmkg = pd.read_csv(annotation_datasets + 'KMKG/full_dataset_KMKG.csv')
-    kmsk = pd.read_csv(annotation_datasets + 'KMSK/full_dataset_KMSK.csv')
-    odile = pd.read_csv(annotation_datasets + 'Odile/full_dataset_Odile.csv')
-    ridim_1 = pd.read_csv(annotation_datasets + 'RIDIM1/full_dataset_RIDIM1.csv')
-    ridim_2 = pd.read_csv(annotation_datasets + 'RIDIM2/full_dataset_RIDIM2.csv')
+    if not os.path.isfile(annotation_datasets + 'FullDataset/' + 'full_dataset_exploration.csv'):
+        flickr = pd.read_csv(annotation_datasets + 'Flickr/full_dataset_Flickr.csv', error_bad_lines=False)
+        kmkg = pd.read_csv(annotation_datasets + 'KMKG/full_dataset_KMKG.csv')
+        kmsk = pd.read_csv(annotation_datasets + 'KMSK/full_dataset_KMSK.csv')
+        odile = pd.read_csv(annotation_datasets + 'Odile/full_dataset_Odile.csv')
+        ridim_1 = pd.read_csv(annotation_datasets + 'RIDIM1/full_dataset_RIDIM1.csv')
+        ridim_2 = pd.read_csv(annotation_datasets + 'RIDIM2/full_dataset_RIDIM2.csv')
 
-    df = pd.concat([flickr, kmkg, kmsk, odile, ridim_1,ridim_2], axis=0)
-    df = df[df.duplicated(subset=["Term"], keep=False)] # Remove instruments which occur only once
-    df.drop_duplicates() # Remove potential duplicates
-    df.to_csv(annotation_datasets + 'FullDataset/' + 'full_dataset_exploration.csv')
+        df = pd.concat([flickr, kmkg, kmsk, odile, ridim_1,ridim_2], axis=0)
+        df = df[df.duplicated(subset=["Term"], keep=False)] # Remove instruments which occur only once
+        df = df.drop_duplicates() # Remove potential duplicates
+        df.to_csv(annotation_datasets + 'FullDataset/' + 'full_dataset_exploration.csv')
 
-    return df
+        print('Stored a full dirty version of the dataset')
+
+    if not os.path.isfile(annotation_datasets + '/FullDataset/MusicArtFullDataset.csv'):
+        IMAGES = '/home/matthia/Documents/Datasets/InstrumentsFullCollection/'
+        DESTINATION = annotation_datasets + 'FullDataset/MusicInArt/'
+        path_regex = '^(/[^/ ]*)+/?$'
+
+        if not os.path.exists(DESTINATION):
+            os.makedirs(DESTINATION)
+
+        df = pd.read_csv(annotation_datasets + 'FullDataset/' + 'full_dataset_exploration.csv')
+        filenames = df['Image Filename']
+        image_list = os.listdir(IMAGES)
+        cnt = 0
+
+        print('Moving and Storing Images')
+
+        for filename in set(filenames):
+            f = os.path.basename(filename)
+            extension = os.path.splitext(filename)[1]
+            for image_name in image_list:
+                if f == image_name:
+                    cnt += 1
+                    df = df.replace(filename,
+                                    'Image_' + str(cnt) + extension)  # replace old file name with new one in df
+                    new_image_name = 'Image_' + str(cnt) + extension  # rename crappy image with a better name
+
+                    shutil.copy(IMAGES + image_name, DESTINATION + new_image_name)  # copy in the final folder
+
+        df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
+        filter = df['Image Filename'].str.contains(path_regex)
+        df = df[~filter]
+
+        df.to_csv(annotation_datasets + 'FullDataset/'+'/MusicArtFullDataset.csv')
+        return df
+
+    else:
+        print('Dataset and Images have already been created')
+        df = pd.read_csv(annotation_datasets + 'FullDataset/'+'/MusicArtFullDataset.csv')
+        df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
+        return df
 
 def explore_training_and_testing_splits(training_df, testing_df):
     """
@@ -106,7 +147,6 @@ def make_splits_tiny_dataset(df):
     """
 
     tiny = df[df['Term'].isin(df['Term'].value_counts()[df['Term'].value_counts() > 25].index)]
-    tiny['Image Filename'] = tiny['Image Filename'].apply(lambda x: x.replace(os.path.dirname(x), dst))
     store_list_of_all_instruments(tiny, 'tiny_version')
 
     le = LabelEncoder()
